+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Unset segment info for web vital spans. ([#6042](https://github.com/getsentry/relay/pull/6042))
 - Set sentry.trace.status on segment spans. ([#6140](https://github.com/getsentry/relay/pull/6140))
 - Don't modify segment information for V2 web vital spans. ([#6160](https://github.com/getsentry/relay/pull/6160))
+- Normalize segment names for standalone spans. ([#6163](https://github.com/getsentry/relay/pull/6163))
 
 **Internal**: